requirements.txt
scrapyd==1.1.0
furl
flask
python-scrapyd-api
pyrfc3339
python-dateutil
commonregex
